refactor: Separate authentication handlers into individual files
