+++ server/handlers/signOut_handler.go
@@ -1,0 +1,16 @@
+package handlers
+
+import (
+	"net/http"
+	"time"
+)
+
+func (h *AuthHandler) SignOut(w http.ResponseWriter, r *http.Request) {
+	http.SetCookie(w, &http.Cookie{
+		Name:    "token",
+		Value:   "",
+		Expires: time.Now().Add(-time.Hour),
+	})
+
+	writeJSONResponse(w, http.StatusOK, JSONResponse{Message: "Logged out"})
+}
