+++ server/handlers/validateToken_handler.go
@@ -1,0 +1,24 @@
+package handlers
+
+import (
+	"log"
+	"net/http"
+)
+
+func (h *AuthHandler) ValidateToken(w http.ResponseWriter, r *http.Request) {
+	tokenString := r.Cookies()[0].Value
+	log.Println(tokenString)
+	if tokenString == "" {
+		writeJSONResponse(w, http.StatusUnauthorized, JSONResponse{Error: MsgNoTokenProvided})
+		return
+	}
+
+	err := h.service.Authenticate(tokenString)
+	if err != nil {
+		log.Println(err)
+		writeJSONResponse(w, http.StatusUnauthorized, JSONResponse{Error: MsgUnauthorized})
+		return
+	}
+
+	writeJSONResponse(w, http.StatusOK, JSONResponse{Message: "Authenticated"})
+}
